Seperate loaders, filters ecc in seperate files to enhance readability
