+++ js/filters.js
@@ -1,97 +1,114 @@
-const filterCuisines = document.getElementById("filterCuisines");
-const filterAuthors = document.getElementById("filterAuthors");
+import { recipes } from "./recipes.js";
+import { loadRecipes } from "./loaders.js";
+const filterArray = [];
 
-const appendClosingButton = () => {
-  filterCuisines.innerHTML += `
-  <button id="close-dropdown">Close</button>
-  `;
-
-  const closingButton = document.getElementById("close-dropdown");
-
-  closingButton.addEventListener("click", () => {
-    filterCuisines.classList.toggle("active");
-  });
+const handleFilterDropdown = (targetId) => {
+  const newTargetId = targetId.charAt(0).toUpperCase() + targetId.slice(1);
+  const filterDropdown = document.getElementById("filter" + newTargetId);
+  filterDropdown.classList.toggle("active");
 };
 
-// Display cuisine
-const displayCuisineCheckboxes = (recipes) => {
-  const cuisineTitle = document.getElementById("cuisine");
+const updateFilters = (e) => {
+  const targetValue = e.target.value;
+  const targetId = e.target.name;
 
-  const getCuisineTypes = () => {
-    // Get all unique cusine types
-    const cuisineSet = new Set();
+  // Check if the targetValue exists in the array
+  const index = filterArray.indexOf(targetValue);
 
-    recipes.forEach((recipe) => {
-      recipe.cuisineType.forEach((cuisineType) => {
-        cuisineSet.add(cuisineType.toLowerCase());
+  // If targetValue exists, remove it; otherwise, push it into the array
+  if (index !== -1) {
+    filterArray.splice(index, 1);
+  } else {
+    filterArray.push(targetValue);
+  }
+
+  if (filterArray.length === 0) {
+    loadRecipes(recipes);
+  } else {
+    const filteredRecipes = recipes.filter((recipe) => {
+      return filterArray.some((value) => {
+        if (recipe.cuisineType && targetId === "cuisine") {
+          return recipe.cuisineType.some(
+            (propValue) => propValue.toLowerCase() === value
+          );
+        } else if (recipe.author && targetId === "author") {
+          return recipe[targetId].toLowerCase() === value;
+        }
+        return false;
       });
     });
+    // Display the filtered recipes
+    loadRecipes(filteredRecipes);
+  }
+};
 
-    // Sort cusine from a -> z
-    return [...cuisineSet].sort();
-  };
+const fetchFilterOptions = (targetId) => {
+  console.log("This is the fetch filter option");
 
-  // Function that will display all the cuisine types in dropdown
-  const createCuisineCheckboxes = (cuisineTypes) => {
-    cuisineTypes.forEach((cuisine) => {
-      filterCuisines.innerHTML += `
-        <label>
-          <input type="checkbox" name="cuisine" class="cuisine" value="${cuisine}"/>
-          ${cuisine}
-        </label> `;
+  const getTargetValues = (targetId) => {
+    console.log("Get all cuisines or authors in the recipe array");
+
+    const targetValues = new Set();
+
+    recipes.forEach((recipe) => {
+      if (targetId === "cuisine") {
+        recipe.cuisineType.forEach((cuisine) =>
+          targetValues.add(cuisine.toLowerCase())
+        );
+      } else if (targetId === "author") {
+        targetValues.add(recipe.author.toLowerCase());
+      }
     });
+    return [...targetValues].sort(); // Convert Set to array and sort
   };
 
-  // Get cuisine types and display cuisine checkboxes
-  const cuisinesTypes = getCuisineTypes();
-  createCuisineCheckboxes(cuisinesTypes);
+  // Function to generate checkboxes based on unique values: Need targetId (cuisine) and the targetValues
+  const generateCheckboxes = (targetId, targetValues) => {
+    console.log("Display checkboxes for the clicked titleTarget");
+    return targetValues
+      .map(
+        (value, index) =>
+          `<label>
+        <input type="checkbox" id="${targetId}-${index}" name="${targetId}" class="${targetId}" value="${value}"/>
+        ${value}
+      </label> `
+      )
+      .join("");
+  };
 
-  // Append closing button
-  appendClosingButton();
+  const targetValues = getTargetValues(targetId);
+  const checkboxesHTML = generateCheckboxes(targetId, targetValues);
 
-  // Eventlistner for cuisine dropdown
-  cuisineTitle.addEventListener("click", () => {
-    filterCuisines.classList.toggle("active");
+  const newTargetId = targetId.charAt(0).toUpperCase() + targetId.slice(1);
+  const filterDropdown = document.getElementById("filter" + newTargetId);
+  filterDropdown.innerHTML = checkboxesHTML;
+
+  // Add event listeners to the checkboxes
+  const checkboxes = filterDropdown.querySelectorAll(
+    `input[type='checkbox'].${targetId}`
+  );
+
+  checkboxes.forEach((checkbox) => {
+    checkbox.addEventListener("change", (e) => {
+      updateFilters(e);
+    });
   });
 };
 
-// Display author
-const displayAuthorCheckboxes = (recipes) => {
-  const authorTitle = document.getElementById("author");
+const filterRecipes = (recipes, e) => {
+  // Get the id from the dropdown
+  const targetId = e.target.id;
 
-  const getAuthorTypes = () => {
-    // Get all unique cusine types
-    const authorSet = new Set();
+  // Toggle dropdown
+  handleFilterDropdown(targetId);
 
-    recipes.forEach((recipe) => {
-      // recipe.source.forEach((authorType) => {
-      authorSet.add(recipe.author.toLowerCase());
-      // });
-    });
+  // Get the checkboxes corresponding to the targetId (cuisine or author)
+  fetchFilterOptions(targetId);
+};
 
-    // Sort cusine from a -> z
-    return [...authorSet].sort();
-  };
-
-  // Function that will display all the cuisine types in dropdown
-  const createAuthorCheckboxes = (authorTypes) => {
-    authorTypes.forEach((author) => {
-      filterAuthors.innerHTML += `
-        <label>
-          <input type="checkbox" name="author" class="author" value="${author}"/>
-          ${author}
-        </label> `;
-    });
-  };
-
-  // Get cuisine types and display cuisine checkboxes
-  const authorTypes = getAuthorTypes();
-  createAuthorCheckboxes(authorTypes);
-  // Append closing button
-  appendClosingButton();
-
-  // Eventlistner for cuisine dropdown
-  authorTitle.addEventListener("click", () => {
-    filterAuthors.classList.toggle("active");
-  });
+export {
+  updateFilters,
+  fetchFilterOptions,
+  filterRecipes,
+  handleFilterDropdown,
 };
